Reorganization of project files. setup.py modification. import reorganization.
