+++ Dockerfile
@@ -5,5 +5,9 @@ LABEL version="0.1"
 LABEL description="Pystrand development dockerfile"
 
 COPY . .
-RUN pip install --no-cache-dir -r requirements.txt
-RUN pip install -e .[tests] --progress-bar off
+
+RUN pip install --no-cache-dir -r requirements.txt \
+    pylint \
+    pytest 
+    
+#RUN pip install -e .[tests] --progress-bar off
