+++ pystrand/__init__.py
@@ -1,0 +1,3 @@
+from pystrand.Genotype import *
+from pystrand.Population import *
+from pystrand.Optimizer import *
