+++ setup.py
@@ -4,7 +4,7 @@ with open("README.md", "r") as fh:
     long_description = fh.read()
 
 setuptools.setup(
-    name="pystrand-pkg-jpodivin", 
+    name="pystrand", 
     version="0.0.1",
     author="Jiri Podivin",
     author_email="jpodivin@gmail.com",
@@ -12,19 +12,20 @@ setuptools.setup(
     long_description=long_description,
     long_description_content_type="text/markdown",
     url="https://github.com/jpodivin/pystrand",
-    packages=setuptools.find_packages(),
+    package_dir={"": "pystrand"},
+    packages=setuptools.find_packages(where="pystrand"),
     classifiers=[
         "Programming Language :: Python :: 3",
         "License :: OSI Approved :: MIT License",
         "Operating System :: OS Independent",
         "Topic :: Scientific/Engineering :: Artificial Life",
     ],
-    python_requires='>=3.6',
+    python_requires=">=3.6",
     install_requires=[ 
-        'numpy'
+        "numpy"
     ],
     extras_require={
-        'tests': ['pytest',
-                  'flake8'],
+        "tests": ["pytest",
+                  "flake8"],
     },
 )
